+++ src/endpoints/ExistsV2.ts
@@ -1,14 +1,17 @@
 import { t } from 'elysia';
 import { AbstractEndpoint } from '../classes/AbstractEndpoint.ts';
-import { DocumentHandler } from '../classes/DocumentHandler.ts';
 import { ErrorHandler } from '../classes/ErrorHandler.ts';
+import { Server } from '../classes/Server.ts';
+import { DocumentUtils } from '../utils/DocumentUtils.ts';
 
 export class ExistsV2 extends AbstractEndpoint {
 	protected override run(): void {
 		this.SERVER.elysia.get(
 			this.PREFIX.concat('/:key/exists'),
 			async ({ params }) => {
-				return DocumentHandler.exists(params);
+				DocumentUtils.validateKey(params.key);
+
+				return Bun.file(Server.DOCUMENT_PATH + params.key).exists();
 			},
 			{
 				params: t.Object({
