+++ src/endpoints/RemoveV1.ts
@@ -1,17 +1,27 @@
+import { unlink } from 'node:fs/promises';
 import { t } from 'elysia';
 import { AbstractEndpoint } from '../classes/AbstractEndpoint.ts';
-import { DocumentHandler } from '../classes/DocumentHandler.ts';
 import { ErrorHandler } from '../classes/ErrorHandler.ts';
+import { Server } from '../classes/Server.ts';
+import { DocumentUtils } from '../utils/DocumentUtils.ts';
 
 export class RemoveV1 extends AbstractEndpoint {
 	protected override run(): void {
 		this.SERVER.elysia.delete(
 			this.PREFIX.concat('/:key'),
 			async ({ headers, params }) => {
-				return DocumentHandler.remove({
-					key: params.key,
-					secret: headers.secret
-				});
+				DocumentUtils.validateKey(params.key);
+
+				const file = await DocumentUtils.retrieveDocument(params.key);
+				const document = await DocumentUtils.documentRead(file);
+
+				DocumentUtils.validateSecret(headers.secret, document.header.modHash);
+
+				return {
+					removed: await unlink(Server.DOCUMENT_PATH + params.key)
+						.then(() => true)
+						.catch(() => false)
+				};
 			},
 			{
 				params: t.Object({
